Move shareable code from Titania to common areas
