+++ service/src/main/kotlin/fi/espoo/evaka/titania/TitaniaQueries.kt
@@ -11,88 +11,6 @@ import fi.espoo.evaka.shared.db.Predicate
 import fi.espoo.evaka.shared.domain.FiniteDateRange
 import fi.espoo.evaka.shared.domain.HelsinkiDateTime
 
-fun Database.Read.employeeNumbersQuery(employeeNumbers: Collection<String>): Database.Query {
-    return createQuery {
-        sql(
-            """
-            SELECT id, employee_number
-            FROM employee
-            WHERE employee_number = ANY (${bind(employeeNumbers)})
-            """
-        )
-    }
-}
-
-fun Database.Read.getEmployeeIdsByNumbers(employeeNumbers: List<String>): Map<String, EmployeeId> {
-    return employeeNumbersQuery(employeeNumbers).toMap { columnPair("employee_number", "id") }
-}
-
-fun Database.Read.getEmployeeIdsByNumbersMapById(
-    employeeNumbers: Collection<String>
-): Map<EmployeeId, String> {
-    return employeeNumbersQuery(employeeNumbers).toMap { columnPair("id", "employee_number") }
-}
-
-fun Database.Read.findStaffAttendancePlansBy(
-    employeeIds: Collection<EmployeeId>? = null,
-    period: FiniteDateRange? = null,
-): List<StaffAttendancePlan> =
-    createQuery {
-            val employeeIdFilter: Predicate =
-                if (employeeIds == null) Predicate.alwaysTrue()
-                else Predicate { where("employee_id = ANY (${bind(employeeIds)})") }
-            val daterangeFilter: Predicate =
-                if (period == null) Predicate.alwaysTrue()
-                else Predicate { where("${bind(period.asHelsinkiDateTimeRange())} @> start_time") }
-
-            sql(
-                """
-            SELECT employee_id, type, start_time, end_time, description
-            FROM staff_attendance_plan
-            WHERE ${predicate(employeeIdFilter.forTable("staff_attendance_plan"))}
-            AND ${predicate(daterangeFilter.forTable("staff_attendance_plan"))}
-        """
-            )
-        }
-        .toList<StaffAttendancePlan>()
-
-fun Database.Transaction.insertStaffAttendancePlans(plans: List<StaffAttendancePlan>): IntArray {
-    if (plans.isEmpty()) {
-        return IntArray(0)
-    }
-    return executeBatch(plans) {
-        sql(
-            """
-INSERT INTO staff_attendance_plan (employee_id, type, start_time, end_time, description)
-VALUES (${bind { it.employeeId }}, ${bind { it.type }}, ${bind { it.startTime }}, ${bind { it.endTime }}, ${bind { it.description }})
-"""
-        )
-    }
-}
-
-fun Database.Transaction.deleteStaffAttendancePlansBy(
-    employeeIds: Collection<EmployeeId>? = null,
-    period: FiniteDateRange? = null,
-): List<StaffAttendancePlan> =
-    createQuery {
-            val employeeIdFilter: Predicate =
-                if (employeeIds == null) Predicate.alwaysTrue()
-                else Predicate { where("employee_id = ANY (${bind(employeeIds)})") }
-            val daterangeFilter: Predicate =
-                if (period == null) Predicate.alwaysTrue()
-                else Predicate { where("${bind(period.asHelsinkiDateTimeRange())} @> start_time") }
-
-            sql(
-                """
-            DELETE FROM staff_attendance_plan
-            WHERE ${predicate(employeeIdFilter.forTable("staff_attendance_plan"))}
-            AND ${predicate(daterangeFilter.forTable("staff_attendance_plan"))}
-            RETURNING employee_id, type, start_time, end_time, description
-        """
-            )
-        }
-        .toList<StaffAttendancePlan>()
-
 fun Database.Read.findStaffAttendancesBy(
     employeeIds: Collection<EmployeeId>? = null,
     period: FiniteDateRange? = null,
